# server/out/public/openApi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/core/database/table/list": {
      "get": {
        "operationId": "DatabaseController_getTableList",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "数据表列表",
        "tags": [
          "数据库管理"
        ]
      }
    },
    "/api/core/database/table/dataSource": {
      "get": {
        "operationId": "DatabaseController_getDataSource",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "数据源信息",
        "tags": [
          "数据库管理"
        ]
      }
    },
    "/api/core/database/table/detailed": {
      "get": {
        "operationId": "DatabaseController_getDetailed",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "表详细信息",
        "tags": [
          "数据库管理"
        ]
      }
    },
    "/api/core/database/recycle/list": {
      "get": {
        "operationId": "DatabaseController_getRecycleList",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "回收站列表",
        "tags": [
          "数据库管理"
        ]
      }
    },
    "/api/core/database/recycle/destroy": {
      "post": {
        "operationId": "DatabaseController_destroy",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "销毁回收站数据",
        "tags": [
          "数据库管理"
        ]
      }
    },
    "/api/core/database/recycle/recovery": {
      "post": {
        "operationId": "DatabaseController_recovery",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "恢复回收站数据",
        "tags": [
          "数据库管理"
        ]
      }
    },
    "/api/core/database/table/optimize": {
      "post": {
        "operationId": "DatabaseController_optimize",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "优化表",
        "tags": [
          "数据库管理"
        ]
      }
    },
    "/api/core/database/table/fragment": {
      "post": {
        "operationId": "DatabaseController_fragment",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "清除表碎片",
        "tags": [
          "数据库管理"
        ]
      }
    },
    "/api/core/database/table/createSql": {
      "get": {
        "operationId": "DatabaseController_getCreateSql",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "查看建表语句",
        "tags": [
          "数据库管理"
        ]
      }
    },
...
    "/api/taixu/document/delete": {
      "post": {
        "operationId": "TaixuDocumentController_delete",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaixuDocumentDeleteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "TaixuDocument"
        ]
      }
    },
...
    "/api/taixu/document/reindex": {
      "post": {
        "operationId": "TaixuDocumentController_reindex",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaixuDocumentReindexDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "TaixuDocument"
        ]
      }
    },
    "/api/taixu/document/website": {
      "post": {
        "operationId": "TaixuDocumentController_website",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaixuDocumentWebsiteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "TaixuDocument"
        ]
      }
    },
...
    "/api/taixu/document/download": {
      "post": {
        "operationId": "TaixuDocumentController_download",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaixuDocumentDownloadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "TaixuDocument"
        ]
      }
    },
...
  },
  "info": {
    "title": "FssAdmin",
    "description": "FssAdmin API 文档",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "Authorization": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "description": "请在请求头中携带 JWT 令牌，格式：Bearer <token>"
      }
    },
    "schemas": {
...
      "TaixuDocumentDeleteDto": {
        "type": "object",
        "properties": {}
      },
      "TaixuDocumentReindexDto": {
        "type": "object",
        "properties": {}
      },
      "TaixuDocumentWebsiteDto": {
        "type": "object",
        "properties": {}
      },
      "TaixuDocumentDownloadDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}